pyproject.toml
[project]
name = "guiacuidarapi"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.7",
    "django-ninja>=1.4.3",
    "supabase>=2.0.0",
    "python-dotenv>=1.0.0",
]
